updated readme and minor updates to example scripts

+++ examples/env_model_function/env_model_function_saasbo.py
@@ -19,7 +19,7 @@ from botorch.models.fully_bayesian import SaasFullyBayesianSingleTaskGP
 from botorch.models.fully_bayesian_multitask import SaasFullyBayesianMultiTaskGP
 from gpytorch.mlls import ExactMarginalLogLikelihood
 
-tkwargs = {"device": torch.device("cpu") if not torch.cuda.is_available() else torch.device("cuda:0"), "dtype": torch.float}
+tkwargs = {"device": torch.device("cpu") if not torch.cuda.is_available() else torch.device("cuda:0"), "dtype": torch.float64}
 
 # Parsing parameters from the command line
 parser = argparse.ArgumentParser()
